horizonclient: Make effect_request examples fully qualified
