+++ clients/horizonclient/examples_test.go
@@ -1,9 +1,12 @@
 package horizonclient_test
 
 import (
+	"context"
 	"fmt"
+	"time"
 
 	"github.com/stellar/go/clients/horizonclient"
+	"github.com/stellar/go/protocols/horizon/effects"
 )
 
 func ExampleClient_NextAssetsPage() {
@@ -35,6 +38,40 @@ func ExampleClient_NextAssetsPage() {
 	fmt.Println(nextPage)
 }
 
+func ExampleClient_NextEffectsPage() {
+	client := horizonclient.DefaultPublicNetClient
+	// all effects
+	effectRequest := horizonclient.EffectRequest{Limit: 20}
+	efp, err := client.Effects(effectRequest)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Print(efp)
+
+	// get next pages.
+	recordsFound := false
+	if len(efp.Embedded.Records) > 0 {
+		recordsFound = true
+	}
+	page := efp
+	// get the next page of records if recordsFound is true
+	for recordsFound {
+		// next page
+		nextPage, err := client.NextEffectsPage(page)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		page = nextPage
+		if len(nextPage.Embedded.Records) == 0 {
+			recordsFound = false
+		}
+		fmt.Println(nextPage)
+	}
+}
+
 func ExampleClient_PrevAssetsPage() {
 	client := horizonclient.DefaultPublicNetClient
 	// assets for asset issuer
@@ -62,4 +99,59 @@ func ExampleClient_PrevAssetsPage() {
 		return
 	}
 	fmt.Println(prevPage)
+}
+
+func ExampleClient_PrevEffectsPage() {
+	client := horizonclient.DefaultPublicNetClient
+	// all effects
+	effectRequest := horizonclient.EffectRequest{Limit: 20}
+	efp, err := client.Effects(effectRequest)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Print(efp)
+
+	// get prev pages.
+	recordsFound := false
+	if len(efp.Embedded.Records) > 0 {
+		recordsFound = true
+	}
+	page := efp
+	// get the prev page of records if recordsFound is true
+	for recordsFound {
+		// prev page
+		prevPage, err := client.PrevEffectsPage(page)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		page = prevPage
+		if len(prevPage.Embedded.Records) == 0 {
+			recordsFound = false
+		}
+		fmt.Println(prevPage)
+	}
+}
+
+func ExampleClient_StreamEffects() {
+	client := horizonclient.DefaultTestNetClient
+	// all effects
+	effectRequest := horizonclient.EffectRequest{Cursor: "760209215489"}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	go func() {
+		// Stop streaming after 60 seconds.
+		time.Sleep(60 * time.Second)
+		cancel()
+	}()
+
+	printHandler := func(e effects.Effect) {
+		fmt.Println(e)
+	}
+	err := client.StreamEffects(ctx, effectRequest, printHandler)
+	if err != nil {
+		fmt.Println(err)
+	}
 }
